+++ CHANGELOG.md
@@ -5,6 +5,7 @@ _Work in progress._
 ### Bug fixes
 - Remove initial blank line(s) in notification emails.
 - Remove invalid expiration date for `github/oauth2` identities.
+- Fix job pagination in evcli.
 
 ## v0.9.0
 First public release.